[1760]: Switch the exchange GetSigners test over to the new RunGetSignersTests helper.
